Eva 72 error edicion cadena aprobacion

See merge request Oficina/eva!65

+++ SGI/views/cadena_aprobacion.py
@@ -34,11 +34,18 @@ class CadenaAprobacionView(AbstractEvaLoggedView):
 
 
 def usuarios_seleccionados(request):
+    """
+    Retorna las selecciones de los usuarios en el formulario de crear-editar cadena de aprobación.
+    Las selecciones se dividen en dos grupos, el primer grupo se construye con JS y se almacenan en el input
+    usuarios_seleccionados, y el segundo grupo se extrae del input ultimo_usuario.
+    :exception: Si no existen valores en el grupo construido por el js, selecciones será igual al unico usuario
+    seleccionado.
+    """
     selecciones = request.POST.get('usuarios_seleccionados', '').split(',')
     if selecciones[0]:
         selecciones.append(request.POST.get('ultimo_usuario', '')[0])
     else:
-        selecciones = request.POST.get('ultimo_usuario', '')
+        selecciones = [request.POST.get('ultimo_usuario', '')]
 
     return selecciones
 
@@ -90,16 +97,16 @@ class CadenaAprobacionEditarView(AbstractEvaLoggedView):
             return render(request, 'SGI/CadenasAprobacion/crear-editar.html',
                           datos_xa_render(self.OPCION, request, cadena, errores))
 
-        if not Documento.objects.filter(cadena_aprobacion_id=id):
+        if not Archivo.objects.filter(cadena_aprobacion_id=id) and \
+                not Documento.objects.filter(cadena_aprobacion_id=id):
             if usuarios_seleccionados:
                 CadenaAprobacionDetalle.objects.filter(cadena_aprobacion=cadena).delete()
+            orden = 1
+            for usuarios in selecciones:
+                CadenaAprobacionDetalle.objects.create(cadena_aprobacion=cadena, usuario_id=usuarios, orden=orden)
+                orden += 1
 
         cadena.save(update_fields=['nombre', 'estado'])
-        orden = 1
-        for usuarios in selecciones:
-            CadenaAprobacionDetalle.objects.create(cadena_aprobacion=cadena, usuario_id=usuarios, orden=orden)
-            orden += 1
-
         messages.success(request, 'Se ha actualizado la cadena de aprobación {0}'.format(cadena.nombre))
         return redirect(reverse('SGI:cadenas-aprobacion-ver'))
